Add button to open integrity report
Fucking hell I hate UI code so much. I know this is a shitty way  to handle adding the button, but I tried making my own XAML window and XL never loaded it so I forced it in. 

This could probably be cleaned up or made into its own window, but it works and I never want to touch WPF again.