Add basic platform check and enhance registry checks
To combat a handful of users who've got vc++ rutimes installed but still have errors, I've enhanced the registry check to check for more information. 

I've also made the DLL check more verbose.

I've also added a super basic platform check to "CanRunDalamud" to error out if it's not x64 for now, as that's the only supported platform for Dalamud.