Always use Windows-style paths for integrity relative path
It turns out the platform check I had was broken, as it's hardcoded to the compiler's host OS. But is also turns out we don't need to care anyways, because we can just always replace `/` with `\\` and it should be fine. 

I also didn't need to do the replacement multiple times, but in my defense, I worked on this at like 4am.