Copy xlcore's method for frontier fallback
Turns out things explode if you run the launcher with  XL_NOAUTOUPDATE because the frontier fallback didn't work / was old.

While I could have just changed a couple 650s to 710s, I didn't really like that patchinstaller also relies on a hardcoded entry.

So I'm copying the method that xlcore uses to try to fetch from kamori first. And then use a gross fallback, but in a place where we'd only have to change it once.

I'm certain there's a better way to do this, but I'm tired and this was the fastest/easiest I could think of.